Setup scripts for local dev
Setup scripts will assume it's in the root plugin directory of OSD
and places the `scripts` folder into the OSD/scripts folder for
run purposes with the following command:
```
yarn setup
```

This then will allow from the root OSD directory to run
```
yarn assistant:start
```
which will setup up the OpenSearch cluster as expected.

Please note, at the time of writing this the feature/os-assistant
and feature/2.11/os-assistant added the feature for installing
OpenSearch assistant related plugins. Not main or 2.x.

Signed-off-by: Kawika Avilla <kavilla414@gmail.com>